Rename "extras" folders to "models"
