+++ c/datatablemodule.cc
@@ -15,8 +15,8 @@
 #include "expr/join_node.h"
 #include "expr/py_expr.h"
 #include "expr/sort_node.h"
-#include "extras/aggregator.h"
-#include "extras/py_ftrl.h"
+#include "models/aggregator.h"
+#include "models/py_ftrl.h"
 #include "frame/py_frame.h"
 #include "options.h"
 #include "py_column.h"
