+++ c/models/murmurhash.cc
@@ -3,7 +3,7 @@
 // Austin Appleby and placed to the public domain, see
 // https://github.com/aappleby/smhasher
 //------------------------------------------------------------------------------
-#include "extras/murmurhash.h"
+#include "models/murmurhash.h"
 #define BIG_CONSTANT(x) (x##LLU)
 
 /*
