+++ c/models/utils.cc
@@ -21,7 +21,7 @@
 //------------------------------------------------------------------------------
 #include <vector>
 #include <stdio.h>
-#include "extras/utils.h"
+#include "models/utils.h"
 
 
 /*
