+++ c/models/utils.h
@@ -19,18 +19,23 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 // IN THE SOFTWARE.
 //------------------------------------------------------------------------------
-#include "extras/py_validator.h"
+#ifndef dt_MODELS_UTILS_h
+#define dt_MODELS_UTILS_h
 
-// Error messages
-Error py::Validator::error_manager::error_not_positive(PyObject* src,
-                                                       const std::string& name
-) const {
-  return ValueError() << name << " should be positive: " << src;
+
+/*
+* Note: this implementation does not disable this overload for array types,
+* see https://en.cppreference.com/w/cpp/memory/unique_ptr/make_unique
+*/
+template<typename T, typename... Args>
+std::unique_ptr<T> make_unique(Args&&... args) {
+  return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
 }
 
-Error py::Validator::error_manager::error_negative(PyObject* src,
-                                                   const std::string& name
-) const {
-  return ValueError() << name << " should be greater than or equal to zero: "
-                      << src;
-}
+#define PBSTR "||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||"
+#define PBWIDTH 60
+void print_progress(float, int);
+
+void calculate_coprimes(size_t, std::vector<size_t>&);
+
+#endif
