+++ tests/models/test_ftrl.py
